Fix all GCC shadowing and conversion warnings from CI
- Fixed parameter shadowing in chess Move constructor (from/to -> from_pos/to_pos)
- Fixed variable shadowing in pomodoro timer (current_time -> now in lambda)
- Fixed parameter shadowing in dbrepair sqlite callback (stmt -> row)
- Fixed sign conversion warning in subnet_scanner (explicit cast to int)
- Fixed variable shadowing in factory::create_card (factory -> factory_it)
- Fixed parameter shadowing in github host::set_login_host (host -> login)
- Fixed constructor parameter shadowing in deck (renderer -> sdl_renderer)
- Fixed size_t to float conversion in deck layout (explicit static_cast)

All fixes ensure compatibility between Clang (local) and GCC (CI) strict warnings.
Local builds now catch the same -Wshadow and -Wconversion issues as CI.

+++ src/cards/system/subnet_scanner.hpp
@@ -468,7 +468,7 @@ private:
                 // Count netmask bits
                 int cidr = 0;
                 for (unsigned int temp = mask_int; temp; temp >>= 1) {
-                    cidr += temp & 1;
+                    cidr += static_cast<int>(temp & 1U);
                 }
                 
                 std::string subnet = std::format("{}/{}", network, cidr);
